added stuff to info.go
